requirements.txt
Flask==0.10.1
httplib2==0.9.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
python-instagram==1.3.1
simplejson==3.6.5
six==1.9.0
Werkzeug==0.10.4
